Refactor damage control functions into their own files
