+++ server/damageReports/functions/generic.js
@@ -1,0 +1,4 @@
+export default ({ message }) => {
+  if (!message) return "";
+  return message;
+};
